[INLONG-5615][Manager] Move sub source to base class for extension
